refactor: Consolidate benchmark result processing scripts into unified process_results.py
Replaces four separate result processing scripts with a single unified script that
handles all benchmark types (TPC-H, TPC-C, TPC-DS, Sysbench) with auto-detection
and consistent interface.

Changes:
- Add scripts/process_results.py with modular parser architecture
- Remove scripts/process_benchmark_results.py (TPC-H)
- Remove scripts/process_tpcc_results.py (TPC-C)
- Remove scripts/process_tpcds_results.py (TPC-DS)
- Remove scripts/process_sysbench_results.py (Sysbench)
- Update scripts/bench to use unified script

Benefits:
- Single entry point for result processing
- ~50% code deduplication through shared utilities
- Auto-detection of benchmark type from content
- Consistent CLI interface across all benchmark types
- Easier to add new benchmark types

Closes #3538
